Move gate expectation regression off live surface
